web/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
supabase==2.3.4
google-generativeai==0.3.2
Pillow==10.2.0
python-magic==0.4.27
django-filter==23.5
django-storages==1.14.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.34
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1 